app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datetime-truncate==1.1.1
Django==5.1.4
django-admin-charts==1.4.0
django-bower==5.2.0
django-environ==0.11.2
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django_csp==3.8
environ==1.0
et_xmlfile==2.0.0
ffmpeg-downloader==0.3.0
ffmpeg-python==0.2.0
firestore-wrapper==0.3.4
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
future==1.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
MouseInfo==0.1.3
multidict==6.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pydub==0.25.1
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytweening==1.2.0
pywhatkit==5.4
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
rsa==4.9
schedule==1.2.2
schema==0.7.7
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.67.1
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whatsapp-wrapper==0.1.6
wikipedia==1.4.0
yarl==1.18.3
